LOCOBOT_TRACKER: Added 'publish_frequecy' node parameter.
LOCOBOT_CONTROL: Added node parameters to get moveit interfaces' names and made some functions' arguments optional so that the class che be inherithed by another class or used as a normal class in a main function, without the need to load parameters. Added an overload function of 'ExecutePlan' for the gripper planning and execution.

ARM_SLEEP_POSITION: Added a cycle that waits the arm is stopped before shutting down the node.