TF_PUBLISHER: Node that publish repetitively a tf passed as parameter. It can be used to publish static tf more than once.
HUMAN_EMULATOR: A node that publish a tf representing the human. This tf changes the pose in the space over the time.

LAUNCH/SIMULATION_NO_LIDAR: Removed the node that was sending the arm to sleep position before the launch file finished. Removed the unused node 'DepthtoScan'.

LAUNCH/DEMO_NO_LIDAR: Changed the executable. Now it launches the state machine node.

STATE_MACHINE: Added services to control the gripper, ask the last error message and clear the error status. The state machine now is able to follow the human along its path and the navigation ends when the robot reaches the human. The target is set 0,5m before the human via nav2 behavior tree 'behavior_tree/NavigationWith_Replanning.xml'. After the navigation is completed the arm is rised and waits for the service to control the gripper to be activated. After that it opens the gripper, waits 10 seconds, closes the gripper and set the arm in sleep position. At this point the action goal return success. In any other case the state machine is sent in error state or aborted if an abort request is made.