Generalized the 'LocobotTracker' class into 'TFRemapper'. Since from now the map frame will be published as a static tf from the camera, there will be no need to locate the map_tag, altough the model will be left in the models' folder. So the specific remapper of the locobot now is just a tf republisher on a desidered topic.
Added launch file to launch the interbotix node and Kobuki base ('robot.launch.py') on the NUC, and the 'remote.launch.py' file to launch the aptiltag recognition from the kinect. Need to evaluate if the kinect launch node should be brought inside this launch file.

On the gesture recognizer has been added a parameter to lower the minimum threshold score for a gesture to be sent.

Added an .rviz file used in the various launch file.

Removed CMakeLists old executable.