In 'LocobotControl' some new access functions has been added and a new error trigger logic has been implemented for the navigation.
In 'StateMachine' an abort state has been added and can be reached from any other state except from error. Furthermore the action server to start the state machine has been removed. Now the the main loop speens in a different thread just before the constructor of the object ends. The end of the separated thread is set by the deconstructor of the object. A mutex has been added to have a safe write/read of the variables used to change state.