Switched to composable nodes for the apriltag detection. This will prevent the publishment of the kinect image to the network, using unnecessary bandwidth.
Added submodules to get the apriltag pose.

With this new 'apriltag_ros' repo, the orientation of the tag is the same of the apriltag library, so it has to be rotated. This is done using a for cycle in the 'remote.launch.py' that publish the rotated pose.

The config file has to be changed to a new format.